+++ src/app/app-routing.module.ts
@@ -34,11 +34,6 @@ const routes: Routes = [
     component: CartListComponent
   },
   {
-    path: '',
-    redirectTo: '/home',
-    pathMatch: 'full'
-  },
-  {
     // The router will match this route if the URL requested
     // doesn't match any paths for routes defined in our configuration
     path: '**',